Module interact::access

Modules

derive

iter

vec

Structs

Function

ImmutAccess
ImmutAccess adds function call information over Reflect.

MutAccess
MutAccess adds function call information over ReflectMut.

Enums

AssignError

CallError

Reflect
An arbitrar between the two possible way to climb into an immutable value.

ReflectMut
An arbitrar between the two possible way to climb into a mutable value.

Traits

Access
The Access trait, meant to be used as a trait object, provides methods that dynamically expose read&write access to the underlying objects.

ReflectDirect
The direct Reflect allows direct climber or reflector access, and meant to be used as a trait object for that purpose.

ReflectIndirect
The indirect Reflect allows indirect climber or reflector access, and meant to be used as a trait object for that purpose.

Functions

deser_assign
A helper for the specific implementations of Access to use with mut_assign methods

Type Definitions

RetValCallback