Trait interact::access::ReflectDirect

pub trait ReflectDirect {
    fn immut_reflector(&self, _reflector: &Arc<Reflector>) -> NodeTree;

    fn immut_climber<'a>(
        &self, 
        _climber: &mut Climber<'a>
    ) -> Result<Option<NodeTree>, ClimbError>;

    fn mut_climber<'a>(
        &mut self, 
        _climber: &mut Climber<'a>
    ) -> Result<Option<NodeTree>, ClimbError>;
}

The direct Reflect allows direct climber or reflector access, and meant to be used as a trait object for that purpose.

Required Methods

fn immut_reflector(&self, _reflector: &Arc<Reflector>) -> NodeTree

The specific implementation of the following method will mostly likely call Reflector::reflect with the specific type.

fn immut_climber<'a>(
&self,
_climber: &mut Climber<'a>
) -> Result<Option<NodeTree>, ClimbError>

Implement climbing for the specific type. Returns a reflection of the inner value, depending on the expression remaining to parse.

fn mut_climber<'a>(
&mut self,
_climber: &mut Climber<'a>
) -> Result<Option<NodeTree>, ClimbError>

Implement mutable climbing for the specific type, allowing to modifying it. Returns a reflection of the inner value, depending on the expression remaining to parse.